refactor/restart improvements
