+++ controller/agent-controller.js
@@ -1,25 +1,32 @@
 import { spawn } from 'child_process';
 
-class AgentController {
+export class AgentController {
     constructor(name) {
         this.name = name;
     }
-    async start(restart_memory=false) {
-        let args = ['init_agent.js', this.name];
+    async start(restart_memory=true) {
+        let args = ['controller/init-agent.js', this.name];
         if (restart_memory)
             args.push('-r');
+
         const agentProcess = spawn('node', args, {
             stdio: 'inherit',
             stderr: 'inherit',
         });
-    
+        
+        let last_restart = Date.now();
         agentProcess.on('exit', (code, signal) => {
             console.log(`Agent process exited with code ${code} and signal ${signal}`);
             
-            // Restart the agent if it exited due to an error
             if (code !== 0) {
+                // agent must run for at least 30 seconds before restarting
+                if (Date.now() - last_restart < 30 * 1000) {
+                    console.error('Agent process exited too quickly. Killing entire process. Goodbye.');
+                    process.exit(1);
+                }
                 console.log('Restarting agent...');
-                this.start();
+                this.start(false);
+                last_restart = Date.now();
             }
         });
     
@@ -28,5 +35,3 @@ class AgentController {
         });
     }
 }
-
-new AgentController('andy').start();
