+++ main.js
@@ -1,0 +1,3 @@
+import { AgentController } from './controller/agent-controller.js';
+
+new AgentController('andy').start();
